PersistentBinaryDeque: Quarantine segments which cannot be read
When a segment has a corrupt header or encounters an error reading mark the
segment as quarantined. This means that the PBD will not immediatly delete or
use the segment. If all readers move past the segment then the segment will be
deleted. The segment is being quarantined instead of deleted in case it is the
only copy of that data and it needs to be recovered.